apps: move block volume related operations to a new file
